containers/ollamalitellm/requirements.txt
ollama
openai
pyyaml
flask
requests
flask[async]
click
flask-socketio
flask-cors
litellm
litellm[proxy]
pymongo